Business logic / ViewModels (#8)
* Title wrapper update

* README update

* Repository pattern updates get method of base repository

* Data instance page - data load

* Data instance creation check of unique name fixed

* Fix entry styles

* HowTo page tempalte update

* Basic Configuration GUI structure

* Combo entry updated for observable collection instead of list

* Added data selection with ComboEntry improvements

* Configuration GUI for data instance is ready for data binding

* Input style improvements + added query GUI

* IsEnable dependencies

* Added loading spinner to DataInstance GUI

* Crawler start/cancel

* Crawler progress update

* Crawler page reload data consistency bug fix

* Data file loading

* Open Raw and Delete files functionality

* Added validation for IndexedDocumentDataModel

* NLP transformed from int to long IDs

* Added transaction system to UoW

* Fixed int to long previous change in tests

* Indexate documents process completion

* Index data files extended with timestamp

* Added storage wrapper into NLP for manipulating data files

* Name corrections

* Index timestamp missing bug fix

* Index file selection

* Query preparation

* Command binding fix

* Localization update

* Radio input

* Radio entry bug fix

* Query logic implemented

* Added query visual validation

* localization update

* Data instance view swapping

* created result context

* Basic query result presentation

* Result styles

* Query select input

* Query error feedback

* Found and Total document count

* Created UC and context for data instance configuration

* Renamed QueryDataResult to QueryDataResultView

* Prepared configuration context

* Added ability to make entries read only

* Added init values into the configuration

* Readonly styles + removed unnecessary styles

* Integer input arrow button styles

* Prepared configuration forms for update methods

* Configuration commands connected

* Basic form method processing

* UoW commit method name change

* Added file deletion on form updates

* Added delete data instance method

* Crawler exception bug fix - invalid site

* form toggling bug fix

+++ InformationRetrievalManager/InformationRetrievalManager.NLP/InvertedIndex/IReadOnlyInvertedIndex.cs
@@ -22,6 +22,6 @@ namespace InformationRetrievalManager.NLP
         /// <remarks>
         ///     <see cref="InvertedIndex._vocabulary"/>
         /// </remarks>
-        IReadOnlyDictionary<string, IReadOnlyDictionary<int, IReadOnlyTermInfo>> GetReadOnlyVocabulary();
+        IReadOnlyDictionary<string, IReadOnlyDictionary<long, IReadOnlyTermInfo>> GetReadOnlyVocabulary();
     }
 }
